+++ Framework/FileOrganiser.php
@@ -1,202 +1,252 @@
 <?php final class FileOrganiser {
 /**
- * All files are stored outside of the webroot (www directory), so it is the job
- * of the FileOrganiser to copy the required files into the webroot when
- * required. The files may need to be minified and compiled before they are
- * copied.
+ * This class works closely with ClientSideCompiler to ensure that all source
+ * files are stored ouside of the webroot (www directory), but the compiled or
+ * minified versions are copied correctly when required.
+ *
+ * The order of execution is controlled by the Dispatcher.
+ *
+ * 1) go functions are executed. This will trigger any PageTools' clientSide()
+ * function, which adds <script> and <link> elements into the DOM head.
+ * 2) FileOrganiser checks if files are needed to be written to www directory.
+ * 3) FileOrganiser writes files to www directory.
+ * 4) ClientSideCompiler pre-processes any SCSS source files.
+ * 5) If client is compiled, FileOrganiser triggers the last step on the
+ * ClientSideCompiler, minifying and compiling all files together and removing
+ * the originals.
+ *
+ * All of this is only done if there are modifications to the source files since
+ * the modified time within www directory.
  */
 
-// List of wildcards to skip.
-private $_skipWc = array(
-	"*.scss", 
-	"ReadMe.md", 
-);
-private $_config;
+private $_wwwDir;
+private $_cacheFile;
 
-public function __construct($config) {
-	// Don't do anything if there are no changes in the source directories.
-	// if(!$this->changedFiles()) {
-	// 	return;
-	// }
-	$this->_config = $config;
+public function __construct() {
+	$this->_wwwDir = APPROOT . "/www";
+	$this->_cacheFile = "{$this->_wwwDir}/www.cache";
 }
 
 /**
- * Checks to see if any source files have changed since the last public copy.
- * @param string $dir Optional. The directory to check. Defaults to /www.
- * @return bool True if there are changes, false if there are no changes.
+ * In production, cache is *always* valid when the www.cache file is present.
+ * To invalidate, just remove the file.
+ *
+ * In non-production, simply obtains the latest modified file time and compares
+ * it to the www.cache's time.
+ *
+ * @return bool True if the www directory needs refreshing.
  */
-private function changedFiles($dir = "/www") {
-	$dir = APPROOT . $dir;
-	$sourceChanged = false;
+public function checkFiles() {
+	$cacheFileExists = file_exists($this->_cacheFile);
+	if(!$cacheFileExists) {
+		return true;
+	}
+	if(App_Config::isProduction()) {
+		return !$cacheFileExists;
+	}
 
-	$wwwMTime = $this->recursiveMTime("/www");
-	$searchDirArray = array(
-		"Script",
-		"Style",
-	);
-	foreach ($searchDirArray as $searchDir) {
-		$searchDir = APPROOT . "/$searchDir";
-		$dirMTime = $this->recursiveMTime($searchDir);
-		if($dirMTime > $wwwMTime) {
-			$sourceChanged = true;
+	$sourceDirectoryArray = array("Asset", "Script", "Style");
+	$fileMTime = 0;
+
+	// First build up the array of files in the source directories.
+	foreach ($sourceDirectoryArray as $sourceDirectory) {
+		// GTROOT comes first in the array, so that they will be overrided by
+		// any files that have the same name in the APPROOT.
+		$directoryPathArray = array(
+			GTROOT . "/$sourceDirectory",
+			APPROOT . "/$sourceDirectory",
+		);
+
+		foreach($directoryPathArray as $directoryPath) {
+			if(!is_dir($directoryPath)) {
+				continue;
+			}
+
+			foreach ($iterator = new RecursiveIteratorIterator(
+				new RecursiveDirectoryIterator($directoryPath,
+					RecursiveDirectoryIterator::SKIP_DOTS),
+			RecursiveIteratorIterator::SELF_FIRST) as $item) {
+
+				$pathName = $iterator->getPathName();
+				$subPathName = $iterator->getSubPathName();
+
+				// Don't check on hidden files or directories.
+				if(strpos($item->getFileName(), ".") === 0
+				|| $item->isDir()) {
+					continue;
+				}
+
+				$thisFileMTime = filemtime($pathName);
+				if($thisFileMTime > $fileMTime) {
+					$fileMTime = $thisFileMTime;
+				}
+			}
 		}
 	}
 
-	return $sourceChanged;
+
+	$cacheMTime = filemtime($this->_cacheFile);
+	// Returns if the cache is out of date.
+	return $cacheMTime < $fileMTime;
 }
 
 /**
- * Gets the latest modified file's timestamp in a directory structure, 
- * recursively.
- * @param  string $dir The absolute directory path to search.
- * @return int         Timestamp of the latest modified file, or 0 if no files.
+ * Removes all files within any subdirectories in the www directory. Doesn't 
+ * remove any files within the root www directory.
  */
-private function recursiveMTime($dir) {
-	$timestamp = 0;
-	if(!is_dir($dir)) {
-		return 0;
-	}
+public function clean() {
+	$directoryArray = array("Asset", "Script", "Style");
+	foreach($directoryArray as $directory) {
+		$directoryPath = "{$this->_wwwDir}/$directory";
 
-	$iterator = new RecursiveDirectoryIterator($dir,
-		RecursiveIteratorIterator::CHILD_FIRST
-		| FilesystemIterator::SKIP_DOTS
-		| FilesystemIterator::UNIX_PATHS);
-	
-	$fileList = new RecursiveIteratorIterator(
-		$iterator, RecursiveIteratorIterator::SELF_FIRST);
-	foreach ($fileList as $key => $value) {
-		if(!is_file($key)) {
+		if(!is_dir($directoryPath)) {
 			continue;
 		}
-		if($value->getMTime() > $timestamp) {
-			$timestamp = $value->getMTime();
+
+		foreach ($iterator = new RecursiveIteratorIterator(
+			new RecursiveDirectoryIterator($directoryPath,
+				RecursiveDirectoryIterator::SKIP_DOTS),
+		RecursiveIteratorIterator::CHILD_FIRST) as $item) {
+
+			$pathName = $iterator->getPathName();
+			unset($pathName);
 		}
+
+		rmdir($directoryPath);
 	}
-	
-	return $timestamp;
+
+	if(file_exists($this->_cacheFile)) {
+		unset($this->_cacheFile);
+	}
 }
 
 /**
- * Deletes all files within the www directory, apart from the vital Go.php.
- * This function is recursive, so will remove all Assets and files within Style.
+ * Performs the actual copying of resources within the three source directories.
+ * Because PageTools can inject client-side resources in the DOM head, a list
+ * of matching elements could be passed in to be included in the file copying.
  */
-public function removePublicFiles() {
-	$dir = APPROOT . "/www/";
-	if(!is_dir($dir)) {
-		// TODO: Throw proper error here.
-		die("ERROR: Public web root is not a directory.");
+public function update($pageToolElements = array()) {
+	$time = time();
+
+	$directoryArray = array("Asset", "Script", "Style");
+	foreach($directoryArray as $directory) {
+		$sourceDirectoryArray = array(
+			GTROOT . "/$directory",
+			APPROOT . "/$directory",
+		);
+
+		foreach ($sourceDirectoryArray as $sourceDirectory) {
+			if(!is_dir($sourceDirectory)) {
+				continue;
+			}
+			
+			foreach ($iterator = new RecursiveIteratorIterator(
+				new RecursiveDirectoryIterator($sourceDirectory,
+					RecursiveDirectoryIterator::SKIP_DOTS),
+			RecursiveIteratorIterator::SELF_FIRST) as $item) {
+
+				$pathName = $iterator->getPathName();
+				$subPathName = $iterator->getSubPathName();
+				$fileName = $item->getFileName();
+				if($fileName[0] === ".") {
+					continue;
+				}
+
+				$destinationFile = "{$this->_wwwDir}/$directory/$subPathName";
+				$destinationDir = dirname($destinationFile);
+
+				if(!is_dir($destinationDir)) {
+					mkdir($destinationDir, 0775, true);
+				}
+
+				if($item->isDir()) {
+					continue;
+				}
+
+				copy($pathName, $destinationFile);
+			}
+		}
 	}
 
-	$iterator = new RecursiveDirectoryIterator($dir,
-		RecursiveIteratorIterator::CHILD_FIRST
-		| FilesystemIterator::SKIP_DOTS
-		| FilesystemIterator::UNIX_PATHS);
-	$fileList = new RecursiveIteratorIterator($iterator, 
-		RecursiveIteratorIterator::CHILD_FIRST);
-	foreach ($fileList as $key => $value) {
-		if(strstr($value->getFilename(), ".php")) {
+	foreach ($pageToolElements as $pageTool) {
+		$source = null;
+		switch(strtolower($pageTool->tagName)) {
+		case "script":
+			$source = $pageTool->getAttribute("src");
+			break;
+		case "style":
+			$source = $pageTool->getAttribute("href");
+			break;
+		default:
+			break;
+		}
+
+		if(is_null($source)) {
 			continue;
 		}
-		if(is_dir($key)) {
-			rmdir($key);
-			continue;
+		
+		$sourcePath = APPROOT . dirname(PATH) . "/$source";
+		$sourcePath = str_replace("//", "/", $sourcePath);
+		$destinationPath = $this->_wwwDir . dirname(PATH) . "/$source";
+		$destinationPath = str_replace("//", "/", $destinationPath);
+
+		$destimationDirectory = dirname($destinationPath);
+		if(!is_dir($destinationDir)) {
+			mkdir($destinationDir, 0775, true);
 		}
-		unlink($key);
+		copy($sourcePath, $destinationPath);
 	}
+
+	file_put_contents($this->_cacheFile, $time);
+	return $time;
 }
 
-public function copyFilesToPublic($config) {
-	// The order of copying is vital here; some applications can overwrite the
-	// files supplied by GT with their own, in whicch case the application's
-	// version of the file will be preferred.
+/**
+ * Some files will require preprocessing, such as SCSS source files. The source
+ * files will be present in the www directory structure, so this function will
+ * REPLACE the source files.
+ */
+public function process($clientSideCompiler) {
+	$count = 0;
 
-	$wwwDir = APPROOT . "/www";
-	$copyDirArray = array(
-		GTROOT  . "/Style/Img/"		=> $wwwDir . "/Style/Img/",
-		GTROOT  . "/Style/Font/"	=> $wwwDir . "/Font/",
-		APPROOT . "/Style/Img"		=> $wwwDir . "/Style/Img/",
-		APPROOT . "/Style/Font/"	=> $wwwDir . "/Font/",
-		APPROOT . "/Asset/"			=> $wwwDir . "/Asset/",
-	);
-	
-	foreach ($copyDirArray as $source => $dest) {
-		$this->copyFiles($source, $dest);
-	}
+	foreach ($iterator = new RecursiveIteratorIterator(
+		new RecursiveDirectoryIterator("$this->_wwwDir/Style",
+			RecursiveDirectoryIterator::SKIP_DOTS),
+	RecursiveIteratorIterator::SELF_FIRST) as $item) {
 
-	$config = $this->_config;
-	if($config::isClientCompiled()) {
-		if(file_exists(APPROOT . "/Script/Script.min.js")) {
-			copy(APPROOT . "/Script/Script.min.js", 
-				APPROOT . "/www/Script.min.js");
+		$pathName = $iterator->getPathName();
+		$fileName = $item->getFileName();
+		$extension = strtolower($item->getExtension());
+		
+		if($extension !== "scss") {
+			continue;
 		}
-		if(file_exists(APPROOT . "/Style/Style.min.css")) {
-			copy(APPROOT . "/Style/Style.min.css", 
-				APPROOT . "/www/Style.min.css");	
+		
+		if($clientSideCompiler->process($pathName)) {
+			$count++;			
 		}
-	}
-	else {
-		$this->copyFiles(GTROOT  . "/Script/", APPROOT . "/www");
-		$this->copyFiles(GTROOT  . "/Style/",  APPROOT . "/www");
-		$this->copyFiles(APPROOT . "/Script/", APPROOT . "/www");
-		$this->copyFiles(APPROOT . "/Style/",  APPROOT . "/www");
-	}
 
+	}
+	return $count;
+}
+
+/**
+ * If client-side compilation is turned on in the App_Config file, this function
+ * compiles all client-side resources into a single resource and DELETES the
+ * original files from the www directory. Note that if the original files do not
+ * change, and the dom head stays the same, the check() function will not allow
+ * this CPU-intensive function to be fired.
+ */
+public function compile($clientSideCompiler, $domHead, 
+$combineForce = false, $compileForce = false) {
+	$isCompiled = App_Config::isClientCompiled();
+	if($isCompiled || $combineForce) {
+		$clientSideCompiler->combine($domHead);
+	}
+	if($isCompiled || $compileForce) {
+		$clientSideCompiler->compile();	
+	}
 	return;
-}
-
-/**
- * Copies the contents of source directory to destination, skipping optional
- * file extensions.
- * @param  string  $source    Source directory to copy from.
- * @param  string  $dest      Destination directory to copy to.
- * @param  boolean $recursive Optional. Whether to perform a deep copy. Defaults
- * to true.
- */
-private function copyFiles($source, $dest, $recursive = true) {
-	if(!is_dir($source)) {
-		return;
-	}
-
-	$dh = opendir($source);
-	if(!is_dir($dest)) {
-		mkdir($dest, 0775, true);
-	}
-
-	while(false !== ($name = readdir($dh)) ) {
-		if($name[0] == ".") {
-			continue;
-		}
-
-		$skip = false;
-		foreach ($this->_skipWc as $skipWc) {
-			if(fnmatch($skipWc, $name)) {
-				$skip = true;
-				break;
-			}
-		}
-		if($skip) {
-			continue;
-		}
-
-		if(is_dir("$source/$name")) {
-			if(!$recursive) {
-				continue;
-			}
-			if(is_dir("$dest/$name")) {
-				continue;
-			}
-			mkdir("$dest/$name", 0775, true);
-			$this->copyFiles(
-				"$source/$name",
-				"$dest/$name",
-				$skipWc,
-				true);
-		}
-		else {
-			copy("$source/$name", "$dest/$name");
-		}
-	}
 }
 
 }#